+++ game-borderlands/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 - None Planned
 
+## [0.2.0] - 2025-09-21
+
+- Autodownload of Python SDK (enables modding and plugin loading).
+- Fixed pathfinding for config and save folders.
+- Added ignoreConflicts list for common files (i.e. LICENSE.txt, instructions.txt, readme.txt, etc.).
+
 ## [0.1.1] - 2025-09-19
 
 - Added installer for movies (.bik files).